+++ src/lib/systems/specials/hull.ts
@@ -1,5 +1,6 @@
-import type { FullThrustShip } from "../../../schemas/ship";
 import { SpecialSystem } from "../_base";
+import type { FullThrustShip } from "../../../schemas/ship";
+import { getSystem } from "..";
 
 export class Hull extends SpecialSystem {
     constructor(ship: FullThrustShip) {
@@ -20,5 +21,20 @@ export class Hull extends SpecialSystem {
         } else if (this.ship.hull.rows === 6) {
             return this.ship.hull.points;
         }
+    }
+
+    cpv() {
+        let ncmass = 0;
+        const ncs: string[] = ["bay", "hangar", "launchTube"];
+        for (const s of this.ship.systems) {
+            if (ncs.includes(s.name)) {
+                const obj = getSystem(s, this.ship);
+                ncmass += obj.mass();
+            }
+        }
+        const realMass = this.ship.mass - ncmass;
+        let cpv = Math.round((realMass * realMass) / 100);
+        if (cpv < 1) { cpv = 1; }
+        return cpv;
     }
 }